backend/requirements.txt
alembic==1.16.2
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cors==1.0.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
requests-file==2.1.0
setuptools==80.9.0
SQLAlchemy==2.0.41
tldextract==5.3.0
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==5.1
zope.interface==7.2
